+++ sickrage/subtitles/__init__.py
@@ -160,6 +160,7 @@ def download_subtitles(episode):
             sickrage.app.log.debug(
                 'history.logSubtitle %s, %s' % (subtitle.provider_name, subtitle.language.opensubtitles))
             History.logSubtitle(episode.show.indexerid,
+                                episode.indexerid,
                                 episode.season,
                                 episode.episode,
                                 episode.status,